refactor: Update mock handlers in tests to use data from separate file
This commit refactors the mock handlers in the tests to use data from a separate file instead of hardcoding the data directly in the handlers. The data is now imported from the "data" file, which contains the categories and products arrays. This improves code maintainability and makes it easier to update the data in the future.